python-fixer/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
Pillow==10.1.0
pdf2docx==0.5.8
lxml
